+++ static_report/src/components/shared/ColumnCard/ColumnTypeDetail/ColumnTypeDetailDatetime.tsx
@@ -1,18 +1,18 @@
 import { Divider, Flex } from '@chakra-ui/react';
 import { ColumnSchema } from '../../../../sdlc/single-report-schema';
-import { DataCompositionMetrics } from '../ColumnMetrics/DataCompositionMetrics';
-import { StatisticalMetrics } from '../ColumnMetrics/StatisticalMetrics';
-import { UniquenessMetrics } from '../ColumnMetrics/UniquenessMetrics';
+import { DataCompositionMatrix } from '../ColumnMatrices/DataCompositionMatrix';
+import { StatisticalMatrix } from '../ColumnMatrices/StatisticalMatrix';
+import { UniquenessMetrics } from '../ColumnMatrices/UniquenessMatrix';
 
 type Props = { columnDatum: ColumnSchema };
 export const ColumnTypeDetailDatetime: React.FC<Props> = ({ columnDatum }) => {
   return (
     <Flex direction={'column'}>
-      <DataCompositionMetrics columnDatum={columnDatum} />
+      <DataCompositionMatrix columnDatum={columnDatum} />
       <Divider />
       <UniquenessMetrics columnDatum={columnDatum} />
       <Divider />
-      <StatisticalMetrics columnDatum={columnDatum} />
+      <StatisticalMatrix columnDatum={columnDatum} />
     </Flex>
   );
 };
