+++ static_report/src/components/shared/ColumnDetailList/ColumnDetailListItem.tsx
