+++ static_report/src/components/shared/ColumnMetrics/SRGeneralStats.tsx
